microservices-assignment/services/order-service/requirements.txt
fastapi
uvicorn[standard]
starlette
psycopg2-binary
python-dotenv
